+++ lib/core/event_builder/index.js
@@ -3,6 +3,8 @@ var environmentUtils = require('../../utils/environment');
 var packageJSON = require('../../../package.json');
 var sprintf = require('sprintf');
 
+var classicOptimizelyEventBuilder = require('./event_builder_classic_optimizely');
+var newOptimizelyEventBuilder = require('./event_builder_new_optimizely');
 var projectConfig = require('../project_config');
 
 var POST_METHOD = 'POST';
@@ -59,12 +61,12 @@ function getCommonEventParams(configObj, userId, attributes) {
     return commonParams;
   } else {
     var params = {};
-    params[urlParams.projectId] = configObj.projectId;
-    params[urlParams.accountId] = configObj.accountId;
-    params[urlParams.endUserId] = userId;
+    params[classicOptimizelyEventBuilder.urlParams.projectId] = configObj.projectId;
+    params[classicOptimizelyEventBuilder.urlParams.accountId] = configObj.accountId;
+    params[classicOptimizelyEventBuilder.urlParams.endUserId] = userId;
 
     if (attributes) {
-      var attributeParams = getAttributeParamsClassicOptimizely(configObj, attributes);
+      var attributeParams = classicOptimizelyEventBuilder.getAttributeParams(configObj, attributes);
       _.assignIn(params, attributeParams);
     }
 
@@ -72,200 +74,10 @@ function getCommonEventParams(configObj, userId, attributes) {
     // executing in
     var sdkName = environmentUtils.isNode() ? 'node-sdk' : 'javascript-sdk';
 
-    params[urlParams.source] = sprintf('%s-%s', sdkName, packageJSON.version);
-    params[urlParams.time] = Math.round(new Date().getTime() / 1000.0);
+    params[classicOptimizelyEventBuilder.urlParams.source] = sprintf('%s-%s', sdkName, packageJSON.version);
+    params[classicOptimizelyEventBuilder.urlParams.time] = Math.round(new Date().getTime() / 1000.0);
     return params;
   }
-}
-
-/*****************************
- * Classic Optimizely endpoint
- *****************************/
-
-var urlParams = {
-  accountId: 'd',
-  projectId: 'a',
-  experimentPrefix: 'x',
-  goalId: 'g',
-  goalName: 'n',
-  segmentPrefix: 's',
-  endUserId: 'u',
-  eventValue: 'v',
-  source: 'src',
-  time: 'time'
-};
-
-/**
- * Get attribute(s) information to the event
- * @param {Object} configObj  Object representing project configuration
- * @param {Object} attributes Object representing user attributes and values which need to be recorded
- */
-function getAttributeParamsClassicOptimizely(configObj, attributes) {
-  var attributeParams = {};
-  _.forEach(attributes, function(attributeValue, attributeKey) {
-    if (attributeValue || attributeValue === false || attributeValue === 0) {
-      var segmentId = configObj.attributeKeyMap[attributeKey].segmentId;
-      var segmentParam = sprintf('%s%s', urlParams.segmentPrefix, segmentId);
-      attributeParams[segmentParam] = attributeValue;
-    }
-  });
-  return attributeParams;
-}
-
-/**
- * Get impression goal information to the event
- * @param  {Object} configObj     Object representing project configuration
- * @param  {string} experimentKey Experiment which is being activated
- * @return {Object}               Impression goal params with impression goal information
- */
-function getImpressionGoalParamsClassicOptimizely(configObj, experimentKey) {
-  var impressionGoalParams = {};
-  var experimentId = projectConfig.getExperimentId(configObj, experimentKey);
-  // For tracking impressions, goal ID is set equal to experiment ID of experiment being activated
-  impressionGoalParams[urlParams.goalId] = experimentId;
-  impressionGoalParams[urlParams.goalName] = 'visitor-event';
-  return impressionGoalParams;
-}
-
-/**
- * Get conversion goal information to the event
- * @param  {Object} configObj  Object representing project configuration
- * @param  {string} eventKey   Goal key representing the event which needs to be recorded
- * @param  {number} eventValue Value associated with the event. Can be used to represent revenue in cents.
- * @return {Object}            Conversion goal params with conversion goal information
- */
-function getConversionGoalParamsClassicOptimizely(configObj, eventKey, eventValue) {
-  var conversionGoalParams = {};
-  var goalId = configObj.eventKeyMap[eventKey].id;
-  var eventIds = goalId;
-  if (eventValue) {
-    eventIds = sprintf('%s,%s', goalId, projectConfig.getRevenueGoalId(configObj));
-    conversionGoalParams[urlParams.eventValue] = eventValue;
-  }
-  conversionGoalParams[urlParams.goalId] = eventIds;
-  conversionGoalParams[urlParams.goalName] = eventKey;
-  return conversionGoalParams;
-}
-
-/**
- * Get experiment to variation mapping to the impression event
- * @param  {Object} configObj     Object representing project configuration
- * @param  {string} experimentKey Experiment which is being activated
- * @param  {string} variationId   Experiment which is being activated
- * @return {Object}               Experiment params with experiment to variation mapping for impression events
- */
-function getExperimentParamsClassicOptimizely(configObj, experimentKey, variationId) {
-  var experimentParams = {};
-  var experimentId = projectConfig.getExperimentId(configObj, experimentKey);
-  var experimentParam = sprintf('%s%s', urlParams.experimentPrefix, experimentId);
-  experimentParams[experimentParam] = variationId;
-  return experimentParams;
-}
-
-/**
- * Maps experiment and corresponding variation as parameters to be used in the event tracking call
- * @param  {Object} configObj                 Object representing project configuration
- * @param  {Array} variationIds              Experiment(s) which are being tracked
- * @param  {Array}  validExperimentIdsForGoal Array of valid experiment IDs that are associated with the event being tracked
- * @return {Object}                           Experiment variation params with experiment to variation mapping for conversion events
- */
-function getExperimentVariationParamsClassicOptimizely(configObj, variationIds, validExperimentIdsForGoal) {
-  var experimentVariationParams = {};
-
-  // Get experiment IDs
-  var experimentKeys = _.filter(Object.keys(configObj.experimentKeyMap), function(experimentKey) {
-    var experimentId = configObj.experimentKeyMap[experimentKey].id;
-
-    if (experimentId && (validExperimentIdsForGoal.indexOf(experimentId) !== -1)) {
-      return experimentKey;
-    }
-  });
-
-  _.forEach(experimentKeys, function(experimentKey) {
-    var experimentId = projectConfig.getExperimentId(configObj, experimentKey);
-    var experimentParam = sprintf('%s%s', urlParams.experimentPrefix, experimentId);
-    var variationId = projectConfig.getEventVariationIdFromExperimentKey(configObj, experimentKey, variationIds);
-    experimentVariationParams[experimentParam] = variationId;
-  });
-
-  return experimentVariationParams;
-}
-
-/*****************************
- * New Optimizely endpoint
- *****************************/
-
-var REVENUE_EVENT_METRIC_NAME = 'revenue';
-
-/**
- *
- * @param configObj     Object representing project configuration
- * @param experimentKey Experiment for which impression needs to be recorded
- * @param variationId   ID for variation which would be presented to user
- * @return {Object}     Impression event specific params for New Optimizely endpoint
- */
-function getImpressionEventParamsNewOptimizely(configObj, experimentKey, variationId) {
-  var impressionEventParams = {
-    layerId: projectConfig.getLayerId(configObj, experimentKey),
-    decision: {
-      isLayerHoldback: false,
-      experimentId: projectConfig.getExperimentId(configObj, experimentKey),
-      variationId: variationId,
-    },
-  };
-  return impressionEventParams;
-}
-
-/**
- *
- * @param {Object} configObj         Object representing project configuration
- * @param {string} eventKey          Event key representing the event which needs to be recorded
- * @param {integer} eventValue       Value associated with the event. Can be used to represent revenue in cents
- * @param {Array} variationIds       Experiment variation ID(s) which are being tracked
- * @param validExperimentIdsForEvent Array of valid experiment IDs that are associated with the event key
- * @return {Object}                  Conversion event specific params for New Optimizely endpoint
- */
-function getConversionEventParamsNewOptimizely(configObj, eventKey, eventValue, variationIds, validExperimentIdsForEvent) {
-  var conversionEventParams = {
-    eventEntityId: projectConfig.getEventId(configObj, eventKey),
-    eventName: eventKey,
-    eventFeatures: [],
-    layerStates: [],
-  };
-
-  // Get valid experiment keys for event
-  var experimentKeys = _.filter(Object.keys(configObj.experimentKeyMap), function(experimentKey) {
-    var experimentId = configObj.experimentKeyMap[experimentKey].id;
-
-    return !!experimentId && (validExperimentIdsForEvent.indexOf(experimentId) !== -1);
-  });
-
-  _.forEach(experimentKeys, function(experimentKey) {
-    var experimentId = projectConfig.getExperimentId(configObj, experimentKey);
-    var variationId = projectConfig.getEventVariationIdFromExperimentKey(configObj, experimentKey, variationIds);
-
-    var layerState = {
-      layerId: projectConfig.getLayerId(configObj, experimentKey),
-      decision: {
-        isLayerHoldback: false,
-        experimentId: experimentId,
-        variationId: variationId,
-      },
-      actionTriggered: true,
-    };
-
-    conversionEventParams.layerStates.push(layerState);
-  });
-
-  conversionEventParams.eventMetrics = [];
-  if (eventValue) {
-    var revenueMetric = {
-      name: REVENUE_EVENT_METRIC_NAME,
-      value: eventValue,
-    };
-    conversionEventParams.eventMetrics.push(revenueMetric);
-  }
-  return conversionEventParams;
 }
 
 module.exports = {
@@ -290,13 +102,13 @@ module.exports = {
     if (options.configObj.version === NEW_OPTIMIZELY_VERSION) {
       impressionEvent.url = NEW_OPTIMIZELY_IMPRESSION_ENDPOINT;
 
-      var impressionEventParams = getImpressionEventParamsNewOptimizely(options.configObj, options.experimentKey, options.variationId);
+      var impressionEventParams = newOptimizelyEventBuilder.getImpressionEventParams(options.configObj, options.experimentKey, options.variationId);
       impressionEvent.params = _.assignIn(commonParams, impressionEventParams);
     } else {
       impressionEvent.url = sprintf(CLASSIC_OPTIMIZELY_ENDPOINT, options.configObj.projectId);
 
-      var impressionGoalParams = getImpressionGoalParamsClassicOptimizely(options.configObj, options.experimentKey);
-      var experimentParams = getExperimentParamsClassicOptimizely(options.configObj, options.experimentKey, options.variationId);
+      var impressionGoalParams = classicOptimizelyEventBuilder.getImpressionGoalParams(options.configObj, options.experimentKey);
+      var experimentParams = classicOptimizelyEventBuilder.getExperimentParams(options.configObj, options.experimentKey, options.variationId);
       impressionEvent.params = _.assignIn(commonParams, impressionGoalParams, experimentParams);
     }
 
@@ -325,13 +137,13 @@ module.exports = {
     if (options.configObj.version === NEW_OPTIMIZELY_VERSION) {
       conversionEvent.url = NEW_OPTIMIZELY_CONVERSION_ENDPOINT;
 
-      var conversionEventParams = getConversionEventParamsNewOptimizely(options.configObj, options.eventKey, options.eventValue, options.variationIds, options.validExperimentIdsForGoal);
+      var conversionEventParams = newOptimizelyEventBuilder.getConversionEventParams(options.configObj, options.eventKey, options.eventValue, options.variationIds, options.validExperimentIdsForGoal);
       conversionEvent.params = _.assignIn(commonParams, conversionEventParams);
     } else {
       conversionEvent.url = sprintf(CLASSIC_OPTIMIZELY_ENDPOINT, options.configObj.projectId);
 
-      var conversionGoalParams = getConversionGoalParamsClassicOptimizely(options.configObj, options.eventKey, options.eventValue);
-      var experimentVariationParams = getExperimentVariationParamsClassicOptimizely(options.configObj, options.variationIds, options.validExperimentIdsForGoal);
+      var conversionGoalParams = classicOptimizelyEventBuilder.getConversionGoalParams(options.configObj, options.eventKey, options.eventValue);
+      var experimentVariationParams = classicOptimizelyEventBuilder.getExperimentVariationParams(options.configObj, options.variationIds, options.validExperimentIdsForGoal);
       conversionEvent.params = _.assignIn(commonParams, conversionGoalParams, experimentVariationParams);
     }
